+++ app/routes/authentication_callbacks/slack.js
@@ -1,6 +1,6 @@
 //
 const request = require('request');
-const config = require('../../../config');
+const config = require('../../config');
 const slackOAuthAccessUrl = 'https://slack.com/api/oauth.access';
 
 module.exports.default = (router) => {
