+++ app/routes/events/detail.js
