+++ app/routes/logout.js
